requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-taggit==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298147424/work
PyJWT==2.4.0
pytz==2021.3
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
